Reorganise home page files
