+++ frontend/src/App.js
@@ -2,14 +2,14 @@ import "./App.css";
 import React, { Component } from "react";
 import { MuiThemeProvider, CssBaseline } from "@material-ui/core";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import Home from "./pages/home/Home";
+import Home from "./pages/Home";
 import Login from "./pages/Login";
 import Room from "./pages/Room";
 import GlobalStyles from "./GlobalStyles";
 import theme from "./theme";
 import ErrorPage from "./pages/ErrorPage";
 import SelectQuestion from "./pages/SelectQuestion";
-import NavBarManager from "./pages/home/NavBarManager";
+import NavBarManager from "./components/HomePage/NavBarManager";
 import Register from "./pages/Register";
 
 class App extends Component {
