+++ frontend/src/components/HomePage/Footer.js
