+++ frontend/src/components/HomePage/HeadSection.js
