+++ frontend/src/components/HomePage/ZoomImage.js
