+++ frontend/src/components/HomePage/calculateSpacing.js
