+++ frontend/src/pages/Home.jsx
@@ -1,9 +1,9 @@
 import React, { Fragment } from "react";
 import PropTypes from "prop-types";
-import HeadSection from "./HeadSection";
-import FeatureSection from "./FeatureSection";
-import PricingSection from "./PricingSection";
-import Footer from "./Footer";
+import HeadSection from "../components/HomePage/HeadSection";
+import FeatureSection from "../components/HomePage/FeatureSection";
+import PricingSection from "../components/HomePage/PricingSection";
+import Footer from "../components/HomePage/Footer";
 
 function Home() {
     return (
